+++ roboteam_ai/src/control/ballHandling/RotateAroundBall.h
@@ -5,7 +5,7 @@
 #ifndef ROBOTEAM_AI_ROTATEAROUNDBALL_H
 #define ROBOTEAM_AI_ROTATEAROUNDBALL_H
 
-#include "../positionControllers/RobotCommand.h"
+#include "roboteam_ai/src/control/RobotCommand.h"
 
 namespace rtt {
 namespace ai {
