+++ roboteam_ai/src/control/numTrees/NumTreePosControl.h
@@ -6,7 +6,7 @@
 #define ROBOTEAM_AI_NUMTREEPOSCONTROL_H
 
 #include <roboteam_ai/src/interface/api/Output.h>
-#include "roboteam_ai/src/control/positionControllers/BasicPosControl.h"
+#include "roboteam_ai/src/control/BasicPosControl.h"
 #include "PathPoint.h"
 #include "Collision.h"
