+++ roboteam_ai/src/skills/Attack.cpp
@@ -2,14 +2,14 @@
 #include "Attack.h"
 #include <roboteam_ai/src/world/Field.h>
 #include <roboteam_ai/src/control/numTrees/NumTreePosControl.h>
-#include <roboteam_ai/src/control/positionControllers/BasicPosControl.h>
+#include <roboteam_ai/src/control/BasicPosControl.h>
 #include <roboteam_ai/src/control/ControlUtils.h>
 #include "roboteam_ai/src/coach/OffensiveCoach.h"
 #include <roboteam_ai/src/control/PositionUtils.h>
 #include "Attack.h"
 #include <roboteam_ai/src/world/Field.h>
 #include <roboteam_ai/src/control/numTrees/NumTreePosControl.h>
-#include <roboteam_ai/src/control/positionControllers/BasicPosControl.h>
+#include <roboteam_ai/src/control/BasicPosControl.h>
 #include <roboteam_ai/src/control/ControlUtils.h>
 #include "roboteam_ai/src/coach/OffensiveCoach.h"
