+++ roboteam_ai/src/skills/Attack.h
@@ -5,7 +5,7 @@
 #ifndef ROBOTEAM_AI_ATTACK_H
 #define ROBOTEAM_AI_ATTACK_H
 
-#include <roboteam_ai/src/control/positionControllers/PosController.h>
+#include <roboteam_ai/src/control/PosController.h>
 #include <roboteam_ai/src/control/shotControllers/ShotController.h>
 #include "Skill.h"
