+++ roboteam_ai/src/skills/CoachDefend.h
@@ -6,7 +6,7 @@
 #define ROBOTEAM_AI_COACHDEFEND_H
 
 #include <roboteam_ai/src/control/numTrees/NumTreePosControl.h>
-#include <roboteam_ai/src/control/positionControllers/BasicPosControl.h>
+#include <roboteam_ai/src/control/BasicPosControl.h>
 #include "Skill.h"
 #include "roboteam_ai/src/control/numTrees/NumTreePosControl.h"
