+++ roboteam_ai/src/skills/InterceptBall.h
@@ -6,7 +6,7 @@
 #define ROBOTEAM_AI_INTERCEPTBALL_H
 
 #include <roboteam_ai/src/control/numTrees/NumTreePosControl.h>
-#include <roboteam_ai/src/control/positionControllers/BasicPosControl.h>
+#include <roboteam_ai/src/control/BasicPosControl.h>
 #include "Skill.h"
 
 namespace rtt {
