+++ roboteam_ai/src/skills/InterceptRobot.hpp
@@ -6,7 +6,7 @@
 #define ROBOTEAM_AI_INTERCEPTROBOT_HPP
 
 #include <roboteam_ai/src/control/numTrees/NumTreePosControl.h>
-#include "roboteam_ai/src/control/positionControllers/BasicPosControl.h"
+#include "roboteam_ai/src/control/BasicPosControl.h"
 
 #include "Skill.h"
 namespace rtt{
