+++ roboteam_ai/src/skills/Pass.h
@@ -6,7 +6,7 @@
 #define ROBOTEAM_AI_PASS_H
 
 #include <roboteam_ai/src/control/numTrees/NumTreePosControl.h>
-#include <roboteam_ai/src/control/positionControllers/BasicPosControl.h>
+#include <roboteam_ai/src/control/BasicPosControl.h>
 #include "Skill.h"
 #include <roboteam_ai/src/coach/PassCoach.h>
 #include <roboteam_ai/src/control/PositionUtils.h>
