+++ roboteam_ai/src/skills/ReflectKick.h
@@ -5,7 +5,7 @@
 #ifndef ROBOTEAM_AI_REFLECTKICK_H
 #define ROBOTEAM_AI_REFLECTKICK_H
 
-#include <roboteam_ai/src/control/positionControllers/BasicPosControl.h>
+#include <roboteam_ai/src/control/BasicPosControl.h>
 #include "InterceptBall.h"
 #include "Skill.h"
 #include <roboteam_ai/src/coach/PassCoach.h>
