+++ roboteam_ai/src/skills/ShootFreeKick.h
@@ -5,7 +5,7 @@
 #ifndef ROBOTEAM_AI_SHOOTFREEKICK_H
 #define ROBOTEAM_AI_SHOOTFREEKICK_H
 
-#include <roboteam_ai/src/control/positionControllers/BasicPosControl.h>
+#include <roboteam_ai/src/control/BasicPosControl.h>
 #include <roboteam_ai/src/control/shotControllers/ShotController.h>
 #include "Skill.h"
 #include "../world/Field.h"
