+++ roboteam_ai/src/skills/gotopos/SkillGoToPos.cpp
@@ -3,7 +3,7 @@
 //
 
 #include <roboteam_ai/src/control/numTrees/NumTreePosControl.h>
-#include <roboteam_ai/src/control/positionControllers/BasicPosControl.h>
+#include <roboteam_ai/src/control/BasicPosControl.h>
 #include <roboteam_ai/src/control/ballHandling/BallHandlePosControl.h>
 #include "SkillGoToPos.h"
